Move collectMetadata method as static in MockConditions
